Refactoring, Cleanup, Changes and documentation

+++ Config/SimulationConfig.cs
@@ -4,9 +4,10 @@
     {
         public const int MapWidth = 20;
         public const int MapHeight = 20;
+
         public const int ChunkWidth = 5;
         public const int ChunkHeight = 5;
 
-        public const int CellSize = 25;
+        public const double CellSize = 25;
     }
 }
